+++ src/main/java/com/learning/swing/entity/GameObject.java
@@ -6,17 +6,17 @@ import java.awt.*;
 
 public abstract class GameObject {
 
-    protected int x, y;
-    protected int velX, velY;
+    protected double x, y;
+    protected double velX, velY;
     protected ID id;
 
-    public GameObject(int x, int y, ID id) {
+    public GameObject(double x, double y, ID id) {
         this.x = x;
         this.y = y;
         this.id = id;
     }
 
-    public static int clamp(int var, int min, int max) {
+    public static double clamp(double var, int min, int max) {
         if (var >= max) {
             return max;
         }
@@ -32,31 +32,31 @@ public abstract class GameObject {
 
     public abstract Rectangle getBounds();
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getVelX() {
+    public double getVelX() {
         return velX;
     }
 
-    public void setVelX(int velX) {
+    public void setVelX(double velX) {
         this.velX = velX;
     }
 
-    public int getVelY() {
+    public double getVelY() {
         return velY;
     }
